+++ src/common/sidebar/containers/SidebarTemplate.tsx
@@ -6,7 +6,7 @@ import { useLocation, useNavigate } from 'react-router';
 
 import { Term } from 'lib/fetchers';
 import { useSessionStorage } from 'lib/hooks/storage/useSessionStorage';
-import { getCurrentTerm } from 'lib/utils/terms';
+import { getCurrentTerm } from 'lib/utils';
 
 import { Feedback } from 'common/feedback';
 import { Header } from 'common/header';
