+++ src/lib/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './dates';
 export * from './terms';
 export * from './courses';
+export * from './constants';
